+++ nata-monkey/src/main/java/com/nata/Config.java
@@ -1,0 +1,68 @@
+package com.nata;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-24 10:19
+ */
+public class Config {
+    private String device_id;
+    private String app_name;
+    private String package_name;
+    private String activity_name;
+    private int action_count;
+    private String algorithm;
+
+    public Config(){
+
+    }
+
+    public String getDevice_id() {
+        return device_id;
+    }
+
+    public void setDevice_id(String device_id) {
+        this.device_id = device_id;
+    }
+
+    public String getApp_name() {
+        return app_name;
+    }
+
+    public void setApp_name(String app_name) {
+        this.app_name = app_name;
+    }
+
+    public String getPackage_name() {
+        return package_name;
+    }
+
+    public void setPackage_name(String package_name) {
+        this.package_name = package_name;
+    }
+
+    public String getActivity_name() {
+        return activity_name;
+    }
+
+    public void setActivity_name(String activity_name) {
+        this.activity_name = activity_name;
+    }
+
+    public int getAction_count() {
+        return action_count;
+    }
+
+    public void setAction_count(int action_count) {
+        this.action_count = action_count;
+    }
+
+    public String getAlgorithm() {
+        return algorithm;
+    }
+
+    public void setAlgorithm(String algorithm) {
+        this.algorithm = algorithm;
+    }
+}
+