+++ nata-server/api_router_v1.js
@@ -1,8 +1,10 @@
 var express = require('express');
 var router = express.Router();
 
-var deviceApi= require('./api/v1/api_device');
+var deviceApi= require('./api/v1/api_device.js');
+var recordApi= require('./api/v1/api_record.js');
 
 router.post('/devices',deviceApi.create);
+router.post('/records',recordApi.create);
 
 module.exports = router;